backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
motor==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
email-validator==1.1.3
boto3==1.26.137
prometheus-client==0.16.0
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
schedule==1.2.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
joblib==1.2.0
celery==5.3.4
redis==5.0.1
pydantic==2.5.2
python-dotenv==1.0.0
stripe==7.6.0
httpx==0.25.2
pymongo==4.6.1 